pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "epitopegraph"
version = "0.1.0"
description = "A graph-based toolkit for conformational B-cell epitope prediction using AlphaFold structures"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Raaga Karumanchi" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "torch>=2.0.0",
    "torch-geometric>=2.3.0",
    "biopython>=1.81",
    "requests>=2.28.0",
    "nglview>=3.0.0",
    "matplotlib>=3.5.0",
    "ipywidgets>=8.0.0",
    "jupyter>=1.0.0",
    "captum>=0.6.0",
    "mdanalysis>=2.4.0",
    "scikit-learn>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=epitopegraph" 